+++ src/app/(home)/page.tsx
@@ -2,30 +2,20 @@
 
 import { useEffect, useRef, useState } from 'react';
 import useMGCTotalList from '@/apis/mgcList/useMGCTotalList';
-import { Separator } from '@/components/ui/separator';
-import { useThunderModalStore } from '@/store/thunderModalStore';
-import useInfoWindowPosition from '@/store/useInfoWindowPosition';
 import useSearchInputValueStore from '@/store/useSearchValueStore';
 import { MGCSummary } from '@/types/MGCList';
-import { X } from 'lucide-react';
-import { useRouter } from 'next/navigation';
 import CreateBtn from '../_components/CreateBtn';
 import MGCList from '../_components/MGCList/MGCList';
 import Map from '../_components/Map/Map';
-import InfoWindow from '../_components/infoWindow/InfoWindow';
 import BottomSheet from './_components/BottomSheet';
 import HomeMap from './_components/HomeMap';
 import HomeMapViewer from './_components/HomeMapViewer';
 
 const Home = () => {
-  const router = useRouter();
-
   const [MGCDataList, setMGCDataList] = useState<MGCSummary[]>([]);
   const [open, setOpen] = useState(false);
 
   const mapRef = useRef<HTMLDivElement>(null);
-
-  const { toggleModal } = useThunderModalStore();
 
   const { searchValue } = useSearchInputValueStore();
 
@@ -59,12 +49,6 @@ const Home = () => {
     handleMouseUp();
   };
 
-  const { infoWindowPosition, setInfoWindowPosition } = useInfoWindowPosition();
-
-  const closeInfoWindow = () => {
-    setInfoWindowPosition({ latitude: 0, longitude: 0 });
-  };
-
   return (
     <div className="relative -left-20pxr w-[100vw]">
       <Map mapRef={mapRef}>
@@ -78,51 +62,6 @@ const Home = () => {
           timerRef={timerRef}
           onMouseUp={handleMouseUp}
         />
-        <InfoWindow
-          show={infoWindowPosition.latitude !== 0 && infoWindowPosition.longitude !== 0}
-          position={infoWindowPosition}
-          getNewPosition={() => {}}
-        >
-          <div
-            id="infowindow-content"
-            // TODO: 임시로 위치 top: -52px해둔 것. 인포윈도우 자체적으로 위치가 이동할 수 았도록 수정 필요 [24.03.18]
-            className="bubble-tail relative -top-52pxr flex h-80pxr flex-col rounded-md bg-white shadow-md"
-          >
-            <div className="flex justify-end">
-              <button
-                className="pr-10pxr pt-7pxr"
-                onClick={closeInfoWindow}
-              >
-                <X
-                  width={13}
-                  height={13}
-                  fill="gray"
-                />
-              </button>
-            </div>
-            <div className="flex grow flex-col items-start justify-evenly">
-              <button
-                className="w-full px-10pxr text-sm"
-                onClick={() => {
-                  router.push('/create');
-                  closeInfoWindow();
-                }}
-              >
-                모각코 생성
-              </button>
-              <Separator />
-              <button
-                className="w-full px-10pxr text-sm"
-                onClick={() => {
-                  toggleModal();
-                  closeInfoWindow();
-                }}
-              >
-                ⚡번개 모각코 생성
-              </button>
-            </div>
-          </div>
-        </InfoWindow>
       </Map>
 
       <div className="absolute bottom-0 right-24pxr z-30">
